requirements.txt
discord.py==2.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.32.3
beautifulsoup4==4.12.3
